Fix game exit crash when under ModOrganizer2, even when only Address Library and STR mods are loaded
Under MO2, there's a bug caused when calling RealGetModuleFileNameW on XAudio2_7.dll, it crashes in usvfs. This only happens during the game shutdown paths when the last few threads are exiting. This PR just avoids that bug.

TODO: Further analysis of what is under MO2 USVFS and what is needed.