WIP for naked NPC fixes
First pass, cleanup needed. I thought I was going to have to dribble out dressing to make this work, but currently it redresses ALL of Whiterun (for example) just to see that it doesn't break. It doesn't break. Randomness isn't needed and will be pulled (the random range is 0 right now, so we're testing the harder everyone dresses at once case).  Filtering redressing to those who need it will eliminate the half-second glitch when something the size of Whiterun redresses.

Fallout4 does not compile yet.

When the 2nd player joins the server (and after), the player systems divvy up the Actors for management. Fights over ownership can occur. Resulting race can cause naked NPCs.

When an actor is claimed, start a deadline timer (to allow fights to settle). When the timer fires, make sure actor is dressed.