Repair Nvidia black-screen fix to reenable ENB mod loading.
The ENB mod places a proxy DLL in the SkyrimSE.exe directory to hook the D3D11 functions it needs. But SkyrimTogether.exe launches from a different directory, so the default DLL search path doesn't find it. Check for existence of the DLL and load it if found.

Also fixes a minor invocation bug.