Simplify (and correct) tracking of modified Master behavior.
New to remember the modified MasterBehavior hash, since it is used to return from Beast mode. Previous implementation stored in ActorExtension, but would only get set on the first modified humanoid which might not be the PlayerCharacter.

Similar issue for dragons, remember the modified hash so Actor::IsDragon() test will work for all modded dragons, not just the first one.